// src/domains/service-clients/payment/proto/generated/payment_service.ts
export interface VerifyPaymentRequest {
  provider: Provider;
  stripe?: StripeVerifyRequest | undefined;
  razorpay?: RazorpayVerifyRequest | undefined;
  paypal?: PayPalVerifyRequest | undefined;
}
// ...
export const VerifyPaymentRequest: MessageFns<VerifyPaymentRequest> = {
  encode(
    message: VerifyPaymentRequest,
    writer: BinaryWriter = new BinaryWriter()
  ): BinaryWriter {
    if (message.provider !== 0) {
      writer.uint32(8).int32(message.provider);
    }
    if (message.stripe !== undefined) {
      StripeVerifyRequest.encode(
        message.stripe,
        writer.uint32(18).fork()
      ).join();
    }
    if (message.razorpay !== undefined) {
      RazorpayVerifyRequest.encode(
        message.razorpay,
        writer.uint32(26).fork()
      ).join();
    }
    if (message.paypal !== undefined) {
      PayPalVerifyRequest.encode(
        message.paypal,
        writer.uint32(34).fork()
      ).join();
    }
    return writer;
  },

  decode(
    input: BinaryReader | Uint8Array,
    length?: number
  ): VerifyPaymentRequest {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseVerifyPaymentRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.provider = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.stripe = StripeVerifyRequest.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.razorpay = RazorpayVerifyRequest.decode(
            reader,
            reader.uint32()
          );
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.paypal = PayPalVerifyRequest.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): VerifyPaymentRequest {
    return {
      provider: isSet(object.provider) ? providerFromJSON(object.provider) : 0,
      stripe: isSet(object.stripe)
        ? StripeVerifyRequest.fromJSON(object.stripe)
        : undefined,
      razorpay: isSet(object.razorpay)
        ? RazorpayVerifyRequest.fromJSON(object.razorpay)
        : undefined,
      paypal: isSet(object.paypal)
        ? PayPalVerifyRequest.fromJSON(object.paypal)
        : undefined,
    };
  },

  toJSON(message: VerifyPaymentRequest): unknown {
    const obj: any = {};
    if (message.provider !== 0) {
      obj.provider = providerToJSON(message.provider);
    }
    if (message.stripe !== undefined) {
      obj.stripe = StripeVerifyRequest.toJSON(message.stripe);
    }
    if (message.razorpay !== undefined) {
      obj.razorpay = RazorpayVerifyRequest.toJSON(message.razorpay);
    }
    if (message.paypal !== undefined) {
      obj.paypal = PayPalVerifyRequest.toJSON(message.paypal);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<VerifyPaymentRequest>, I>>(
    base?: I
  ): VerifyPaymentRequest {
    return VerifyPaymentRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<VerifyPaymentRequest>, I>>(
    object: I
  ): VerifyPaymentRequest {
    const message = createBaseVerifyPaymentRequest();
    message.provider = object.provider ?? 0;
    message.stripe =
      object.stripe !== undefined && object.stripe !== null
        ? StripeVerifyRequest.fromPartial(object.stripe)
        : undefined;
    message.razorpay =
      object.razorpay !== undefined && object.razorpay !== null
        ? RazorpayVerifyRequest.fromPartial(object.razorpay)
        : undefined;
    message.paypal =
      object.paypal !== undefined && object.paypal !== null
        ? PayPalVerifyRequest.fromPartial(object.paypal)
        : undefined;
    return message;
  },
};
// ...
type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | boolean
  | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}